Refuse TCP connections if they fail to connect
Prior to this commit, TCP connections that fail to connect will simply hang,
necessitating the client to time out instead of immediately failing.

After this commit, the TCP connection is accepted _only_ after the remote
SOCKS connection is successfully made. If it fails, an early RST is sent,
which is interpreted as Connection refused for the client.