Drop capabilities for main process
We used to only drop capabilities in daemon process, since kernel
clear the capabilities on exec for us.

However, this is only the case for file capabilities.

On some systems(e.g. NixOS), capabilities are granted through a wrapper
program that set up ambient capabilities.

Ambient capabilities are NOT automatically cleared by kernel on
exec. Thereby we clear the capabilities explicitly before exec.