Make --fake-dns a string type flag
It might be confusing, if you run proxy-ns like this:

$ proxy-ns --fake-dns false curl g.co

As the flag package in go doesn't allow boolean flags to be written in
the form like "-flag x", it actually executes the program named
"false" with arguments as ["curl", "g.co"].

So the command above will fail with status code 1.

To avoid this confusion, we use a string type flag for
"--fake-dns", then parse the string value manually later.